chore: Rename components for clarity
Renamed StudentDetails to ManualAttendanceSection
Renamed StudentDetailsDisplay to StudentDetailsDisplay
